hardware-controller/requirements.txt
# GPIO Control (Raspberry Pi)
RPi.GPIO==0.7.1
gpiozero==2.0.1

# Alternative for Orange Pi
# OPi.GPIO==0.5.3

# Redis for command/status communication
redis==5.0.1
hiredis==2.3.2

# Configuration
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Async support
asyncio==3.4.3

# Testing
pytest==7.4.4
